Andromeda-package {Andromeda}	R Documentation

Andromeda: Asynchronous Disk-Based Representation of Massive Data

Description

Storing very large data objects on a local drive, while still making it possible to manipulate the data in an efficient manner.

Author(s)

Maintainer: Adam Black black@ohdsi.org

Authors:

Martijn Schuemie schuemie@ohdsi.org

Marc A. Suchard msuchard@ucla.edu

Other contributors:

Observational Health Data Science and Informatics [copyright holder]

See Also

Useful links:

https://ohdsi.github.io/Andromeda/

https://github.com/OHDSI/Andromeda

Report bugs at https://github.com/OHDSI/Andromeda/issues

[Package Andromeda version 0.4.1 Index]