loadAndromeda {Andromeda}	R Documentation

Load Andromeda from file

Description

Load Andromeda from file

Usage

loadAndromeda(fileName)

Arguments

fileName
The path where the object was saved using saveAndromeda().

Value

An Andromeda object.

See Also

saveAndromeda()

Examples

# For this example we create an Andromeda object and save it to
# a temporary file locationL
fileName <- tempfile()
andr <- andromeda(cars = cars)
saveAndromeda(andr, fileName)

# Using loadAndromeda to load the object back:
andr <- loadAndromeda(fileName)

# Don't forget to close Andromeda when you are done:
close(andr)

# Cleaning up the file used in this example:
unlink(fileName)

[Package Andromeda version 0.4.1 Index]