removeIndex {Andromeda}	R Documentation

Removes an index from an Andromeda table

Description

Removes an index from an Andromeda table

Usage

removeIndex(tbl, columnNames = NULL, indexName = NULL)

Arguments

tbl
An Andromeda table (or any other 'DBI' table).

columnNames
A vector of column names (character) on which the index was created. If not provided, then the indexName argument must be provided.

indexName
The name of the index. If not provided, the columnNames argument must be provided.

Details

Remove an index created using the createIndex() function. Either the index name or the column names on which the index was created must be provided.

Value

Invisibly returns the input table.

See Also

createIndex(), listIndices()

Examples

andr <- andromeda(cars = cars)

createIndex(andr$cars, "speed")

# Will be faster now that speed is indexed:
andr$cars %>%
  filter(speed == 10) %>%
  collect()
  
removeIndex(andr$cars, "speed")

close(andr)

[Package Andromeda version 0.4.1 Index]