Save Andromeda to file

Saves the Andromeda object in a zipped file. Note that by default the Andromeda object is automatically closed by saving it to disk. This is due to a limitation of the underlying technology ('RSQLite'). To keep the connection open, use maintainConnection = TRUE. This will first create a temporary copy of the Andromeda object. Note that this can be substantially slower.

saveAndromeda(
  andromeda,
  fileName,
  maintainConnection = FALSE,
  overwrite = TRUE
)

Arguments

andromeda

An object of class Andromeda.

fileName

The path where the object will be written.

maintainConnection

Should the connection be maintained after saving? If FALSE, the Andromeda object will be invalid after this operation, but saving will be faster.

overwrite

If the file exists, should it be overwritten? If FALSE and the file exists, an error will be thrown.

Value

Returns no value. Executed for the side-effect of saving the object to disk.

See also

Examples

andr <- andromeda(cars = cars)

# For this example we'll use a temporary file location:
fileName <- tempfile()

saveAndromeda(andr, fileName)
#> Disconnected Andromeda. This data object can no longer be used

# Cleaning up the file used in this example:
unlink(fileName)