Create an Andromeda object

By default the Andromeda object is created in the systems temporary file location. You can override this by specifying a folder using options(andromedaTempFolder = "c:/andromedaTemp"), where "c:/andromedaTemp" is the folder to create the Andromeda objects in.

andromeda(...)

Arguments

...

Named objects. See details for what objects are valid. If no objects are provided, an empty Andromeda is returned.

Value

Returns an Andromeda object.

Details

Valid objects are data frames, Andromeda tables, or any other dplyr table.

Examples

andr <- andromeda(cars = cars, iris = iris)

names(andr)
#> [1] "cars" "iris"
# [1] 'cars' 'iris'

andr$cars %>% filter(speed > 10) %>% collect()
#> # A tibble: 41 × 2
#>    speed  dist
#>    <dbl> <dbl>
#>  1    11    17
#>  2    11    28
#>  3    12    14
#>  4    12    20
#>  5    12    24
#>  6    12    28
#>  7    13    26
#>  8    13    34
#>  9    13    34
#> 10    13    46
#> # ℹ 31 more rows
# # A tibble: 41 x 2 
# speed dist 
# <dbl> <dbl> 
# 1 11 17 
# ...

close(andr)