andromeda {Andromeda}	R Documentation

Create an Andromeda object

Description

By default the Andromeda object is created in the systems temporary file location. You can override this by specifying a folder using options(andromedaTempFolder = "c:/andromedaTemp"), where "c:/andromedaTemp" is the folder to create the Andromeda objects in.

Usage

andromeda(...)

Arguments

...
Named objects. See details for what objects are valid. If no objects are provided, an empty Andromeda is returned.

Details

Valid objects are data frames, Andromeda tables, or any other dplyr table.

Value

Returns an Andromeda object.

Examples

andr <- andromeda(cars = cars, iris = iris)

names(andr)
# [1] 'cars' 'iris'

andr$cars %>% filter(speed > 10) %>% collect()
# # A tibble: 41 x 2 
# speed dist 
# <dbl> <dbl> 
# 1 11 17 
# ...

close(andr)

[Package Andromeda version 0.4.1 Index]