List all indices on an Andromeda table

List all indices on an Andromeda table

listIndices(tbl)

Arguments

tbl

An Andromeda table (or any other 'DBI' table).

Value

Returns a tibble listing the indices, indexed columns, and whether the index is unique.

Details

Lists any indices that may have been created using the createIndex() function.

See also

Examples

andr <- andromeda(cars = cars)

createIndex(andr$cars, "speed")

listIndices(andr$cars)
#> # A tibble: 1 × 14
#>   database_name   database_oid schema_name schema_oid index_name       index_oid
#>   <chr>                  <dbl> <chr>            <dbl> <chr>                <dbl>
#> 1 file37399085c73          570 main               572 idx_pwo5t0a1y2j…      1333
#> # ℹ 8 more variables: table_name <chr>, table_oid <dbl>, comment <chr>,
#> #   tags <list>, is_unique <lgl>, is_primary <lgl>, expressions <chr>,
#> #   sql <chr>
# # A tibble: 1 x 5
# indexSequenceId indexName                unique columnSequenceId columnName
#           <int> <chr>                    <lgl>             <int> <chr>     
#1              0 idx_ocy8we9j2i7ld0rshgb4 FALSE                 0 speed           

close(andr)