listIndices {Andromeda}	R Documentation

List all indices on an Andromeda table

Description

List all indices on an Andromeda table

Usage

listIndices(tbl)

Arguments

tbl
An Andromeda table (or any other 'DBI' table).

Details

Lists any indices that may have been created using the createIndex() function.

Value

Returns a tibble listing the indices, indexed columns, and whether the index is unique.

See Also

createIndex(), removeIndex()

Examples

andr <- andromeda(cars = cars)

createIndex(andr$cars, "speed")

listIndices(andr$cars)
# # A tibble: 1 x 5
# indexSequenceId indexName                unique columnSequenceId columnName
#           <int> <chr>                    <lgl>             <int> <chr>     
#1              0 idx_ocy8we9j2i7ld0rshgb4 FALSE                 0 speed           

close(andr)

[Package Andromeda version 0.4.1 Index]